+++ database/install/myems_system_db.sql
@@ -1234,6 +1234,7 @@ VALUES
 (112,'Plan','/space/plan',100,1),
 (113,'Production','/space/production',100,1),
 (114,'Enter Production','/space/enterproduction',100,1),
+(115,'Prediction','/space/prediction',100,1),
 (200,'Equipment Data','/equipment',NULL,0),
 (201,'Energy Category Data','/equipment/energycategory',200,0),
 (202,'Energy Item Data','/equipment/energyitem',200,0),